Renamed 'main' to 'server' to reflect now hierarchy
